adding assignment2 files and minor changes.

+++ Hello/integer_bases.c
@@ -6,8 +6,8 @@
 
 int main(void)
 {
-  int v = 10;
-    printf("the base 10 (decimal) value of v is: %d\n", v);
+  float v = 10;
+    printf("the base 10 (decimal) value of v is: %f\n", v);
     printf("the base 8 (octal) value of v is: %#o\n", v);
     printf("the base 16 (hexadecimal) value of v is: %#x\n", v);
     printf("\n");
